+++ src/interface/swarm_comms.cpp
@@ -25,59 +25,6 @@
 
 namespace parthenon {
 
-template <class BOutflow, class BPeriodic, int iFace>
-void Swarm::AllocateBoundariesImpl_(MeshBlock *pmb) {
-  std::stringstream msg;
-  auto &bcs = pmb->pmy_mesh->mesh_bcs;
-  if (bcs[iFace] == BoundaryFlag::outflow) {
-    bounds_uptrs[iFace] = DeviceAllocate<BOutflow>();
-  } else if (bcs[iFace] == BoundaryFlag::periodic) {
-    bounds_uptrs[iFace] = DeviceAllocate<BPeriodic>();
-  } else if (bcs[iFace] == BoundaryFlag::user) {
-    if (pmb->pmy_mesh->SwarmBndryFnctn[iFace] != nullptr) {
-      bounds_uptrs[iFace] = pmb->pmy_mesh->SwarmBndryFnctn[iFace]();
-    } else {
-      msg << (iFace % 2 == 0 ? "i" : "o") << "x" << iFace / 2 + 1
-          << " user boundary requested but provided function is null!";
-      PARTHENON_THROW(msg);
-    }
-  } else {
-    msg << (iFace % 2 == 0 ? "i" : "o") << "x" << iFace / 2 + 1 << " boundary flag "
-        << static_cast<int>(bcs[iFace]) << " not supported!";
-    PARTHENON_THROW(msg);
-  }
-}
-
-void Swarm::AllocateBoundaries() {
-  auto pmb = GetBlockPointer();
-  std::stringstream msg;
-
-  auto &bcs = pmb->pmy_mesh->mesh_bcs;
-
-  AllocateBoundariesImpl_<ParticleBoundIX1Outflow, ParticleBoundIX1Periodic, 0>(
-      pmb.get());
-  AllocateBoundariesImpl_<ParticleBoundOX1Outflow, ParticleBoundOX1Periodic, 1>(
-      pmb.get());
-  AllocateBoundariesImpl_<ParticleBoundIX2Outflow, ParticleBoundIX2Periodic, 2>(
-      pmb.get());
-  AllocateBoundariesImpl_<ParticleBoundOX2Outflow, ParticleBoundOX2Periodic, 3>(
-      pmb.get());
-  AllocateBoundariesImpl_<ParticleBoundIX3Outflow, ParticleBoundIX3Periodic, 4>(
-      pmb.get());
-  AllocateBoundariesImpl_<ParticleBoundOX3Outflow, ParticleBoundOX3Periodic, 5>(
-      pmb.get());
-
-  for (int n = 0; n < 6; n++) {
-    bounds_d.bounds[n] = bounds_uptrs[n].get();
-    std::stringstream msg;
-    msg << "Boundary condition on face " << n << " missing.\n"
-        << "Please set it to `outflow`, `periodic`, or `user` in the input deck.\n"
-        << "If you set it to user, you must also manually set "
-        << "the swarm boundary pointer in your application." << std::endl;
-    PARTHENON_REQUIRE(bounds_d.bounds[n] != nullptr, msg);
-  }
-}
-
 ///
 /// Routine for precomputing neighbor indices to efficiently compute particle
 /// position in terms of neighbor blocks based on spatial position. See
@@ -380,7 +327,6 @@ void Swarm::SetupPersistentMPI() {
 }
 
 int Swarm::CountParticlesToSend_() {
-  auto block_index_h = block_index_.GetHostMirrorAndCopy();
   auto mask_h = Kokkos::create_mirror_view_and_copy(HostMemSpace(), mask_);
   auto swarm_d = GetDeviceContext();
   auto pmb = GetBlockPointer();
@@ -396,8 +342,23 @@ int Swarm::CountParticlesToSend_() {
   const int particle_size = GetParticleDataSize();
   vbswarm->particle_size = particle_size;
 
+  // TODO(BRR) This kernel launch should be folded into the subsequent logic once we
+  // convert that to kernel-based reductions
+  auto &x = Get<Real>(swarm_position::x::name()).Get();
+  auto &y = Get<Real>(swarm_position::y::name()).Get();
+  auto &z = Get<Real>(swarm_position::z::name()).Get();
+  const int max_active_index = GetMaxActiveIndex();
+  pmb->par_for(
+      PARTHENON_AUTO_LABEL, 0, max_active_index, KOKKOS_LAMBDA(const int n) {
+        if (swarm_d.IsActive(n)) {
+          bool on_current_mesh_block = true;
+          swarm_d.GetNeighborBlockIndex(n, x(n), y(n), z(n), on_current_mesh_block);
+        }
+      });
+
   int max_indices_size = 0;
   int total_noblock_particles = 0;
+  auto block_index_h = block_index_.GetHostMirrorAndCopy();
   for (int n = 0; n <= max_active_index_; n++) {
     if (mask_h(n)) {
       // This particle should be sent
@@ -431,7 +392,6 @@ int Swarm::CountParticlesToSend_() {
       }
     }
     noblock_indices.DeepCopy(noblock_indices_h);
-    ApplyBoundaries_(total_noblock_particles, noblock_indices);
   }
 
   // TODO(BRR) don't allocate dynamically
@@ -518,6 +478,7 @@ void Swarm::Send(BoundaryCommSubset phase) {
   auto swarm_d = GetDeviceContext();
 
   if (nneighbor == 0) {
+    // TODO(BRR) Do we ever reach this branch?
     // Process physical boundary conditions on "sent" particles
     auto block_index_h = block_index_.GetHostMirrorAndCopy();
     auto mask_h = Kokkos::create_mirror_view_and_copy(HostMemSpace(), mask_);
@@ -547,8 +508,6 @@ void Swarm::Send(BoundaryCommSubset phase) {
         }
       }
       new_indices.DeepCopy(new_indices_h);
-
-      ApplyBoundaries_(total_sent_particles, new_indices);
     }
   } else {
     // Query particles for those to be sent
@@ -644,28 +603,7 @@ void Swarm::UnloadBuffers_() {
             bid++;
           }
         });
-
-    ApplyBoundaries_(total_received_particles_, new_indices_);
   }
-}
-
-void Swarm::ApplyBoundaries_(const int nparticles, ParArray1D<int> indices) {
-  auto pmb = GetBlockPointer();
-  auto &x = Get<Real>(swarm_position::x::name()).Get();
-  auto &y = Get<Real>(swarm_position::y::name()).Get();
-  auto &z = Get<Real>(swarm_position::z::name()).Get();
-  auto swarm_d = GetDeviceContext();
-  auto bcs = this->bounds_d;
-
-  pmb->par_for(
-      PARTHENON_AUTO_LABEL, 0, nparticles - 1, KOKKOS_LAMBDA(const int n) {
-        const int sid = indices(n);
-        for (int l = 0; l < 6; l++) {
-          bcs.bounds[l]->Apply(sid, x(sid), y(sid), z(sid), swarm_d);
-        }
-      });
-
-  RemoveMarkedParticles();
 }
 
 bool Swarm::Receive(BoundaryCommSubset phase) {
@@ -675,30 +613,30 @@ bool Swarm::Receive(BoundaryCommSubset phase) {
   if (nneighbor == 0) {
     // Do nothing; no boundaries to receive
     return true;
-  } else {
-    // Ensure all local deep copies marked BoundaryStatus::completed are actually
-    // received
-    pmb->exec_space.fence();
-
-    // Populate buffers
-    vbswarm->Receive(phase);
-
-    // Transfer data from buffers to swarm memory pool
-    UnloadBuffers_();
-
-    auto &bdvar = vbswarm->bd_var_;
-    bool all_boundaries_received = true;
-    for (int n = 0; n < nneighbor; n++) {
-      NeighborBlock &nb = pmb->neighbors[n];
-      if (bdvar.flag[nb.bufid] == BoundaryStatus::arrived) {
-        bdvar.flag[nb.bufid] = BoundaryStatus::completed;
-      } else if (bdvar.flag[nb.bufid] == BoundaryStatus::waiting) {
-        all_boundaries_received = false;
-      }
-    }
-
-    return all_boundaries_received;
   }
+
+  // Ensure all local deep copies marked BoundaryStatus::completed are actually
+  // received
+  pmb->exec_space.fence();
+
+  // Populate buffers
+  vbswarm->Receive(phase);
+
+  // Transfer data from buffers to swarm memory pool
+  UnloadBuffers_();
+
+  auto &bdvar = vbswarm->bd_var_;
+  bool all_boundaries_received = true;
+  for (int n = 0; n < nneighbor; n++) {
+    NeighborBlock &nb = pmb->neighbors[n];
+    if (bdvar.flag[nb.bufid] == BoundaryStatus::arrived) {
+      bdvar.flag[nb.bufid] = BoundaryStatus::completed;
+    } else if (bdvar.flag[nb.bufid] == BoundaryStatus::waiting) {
+      all_boundaries_received = false;
+    }
+  }
+
+  return all_boundaries_received;
 }
 
 void Swarm::ResetCommunication() {